app/requirements.txt
Flask
flask_sqlalchemy
pandas
matplotlib
mysql-connector-python
requests

